refactor name rules
