+++ tests/asr/test_asr_opea_whisper_on_intel_hpu.sh
